test/chaos-pipeline/apps/rest-patch-stops-working-after-restart/requirements.txt
weaviate-client==3.5.1
loguru==0.6.0
